readlsm() relies on freadtokval()
git-svn-id: svn://svn.svardos.org/svardos@249 911cea91-c70f-4353-bd03-772f58fe8c9d

+++ pkginst/lsm.h
@@ -1,10 +1,10 @@
 /*
- * This file is part of FDNPKG
- * Copyright (C) 2013-2016 Mateusz Viste, All rights reserved.
+ * This file is part of pkginst
+ * Copyright (C) 2013-2021 Mateusz Viste, All rights reserved.
  */
 
 #ifndef readlsm_h_sentinel
   #define readlsm_h_sentinel
   /* Loads metadata from an LSM file. Returns 0 on success, non-zero on error. */
-  int readlsm(const char *filename, char *version, int version_maxlen);
+  int readlsm(const char *filename, char *version, size_t version_maxlen);
 #endif
